deploy_tool/requirements.txt
click>=8.0.0
boto3>=1.18.0
gitpython>=3.1.20
cryptography>=38.0.0
